Fix segmentation fault in hammer
If an OWRITE is attempted but fails due to HAMMER_DISK_FULL and the hammer_file_t is freefile()'d,
further attempts to use the hammer_file_t structure in hammer_doio caused a seg fault.

This change fixes the issue by setting the hammer_file_t pointer to NULL so further nullity checks correctly get the state of the file data structure.